Refactor into modular structure
