+++ main.py
@@ -24,418 +24,17 @@ import pandas as pd
 from openpyxl import Workbook, load_workbook
 from openpyxl.styles import Font, PatternFill, Border, Side, Alignment, NamedStyle
 from openpyxl.utils import get_column_letter
-from openpyxl.styles import Font, PatternFill, Border, Side, Alignment, NamedStyle
 from openpyxl.formatting.rule import ColorScaleRule, DataBarRule
 from openpyxl.chart import BarChart, LineChart, Reference
 from openpyxl.utils import get_column_letter
 
+from excel_styler import ExcelStyler
+from license_manager import LicenseManager, LicenseDialog
+from console_widget import ConsoleWidget
+
 # Configurar tema de customtkinter
 ctk.set_appearance_mode("dark")
 ctk.set_default_color_theme("blue")
-
-class ExcelStyler:
-    """Clase para manejar estilos del Excel de forma profesional"""
-    
-    def __init__(self):
-        # Definir paleta de colores corporativa
-        self.colors = {
-            'primary': 'FF2B5CE6',      # Azul principal
-            'secondary': 'FF10B981',    # Verde éxito  
-            'accent': 'FFFBBF24',       # Amarillo acento
-            'danger': 'FFEF4444',       # Rojo error
-            'dark': 'FF1F2937',         # Gris oscuro
-            'light': 'FFF8FAFC',        # Gris claro
-            'white': 'FFFFFFFF',        # Blanco
-            'header_bg': 'FF1E40AF',    # Azul header
-            'alt_row': 'FFF1F5F9',      # Fila alternativa
-            'neto_positive': 'FF10B981', # Verde para neto positivo
-            'neto_negative': 'FFEF4444', # Rojo para neto negativo
-            'haberes_bg': 'FF10B981',   # Verde para sección haberes
-            'deducciones_bg': 'FFEF4444', # Rojo para sección deducciones
-            'totales_bg': 'FF6366F1'    # Púrpura para totales
-        }
-    
-    def create_header_style(self):
-        """Estilo para headers de tabla"""
-        header_style = NamedStyle(name="header_style")
-        header_style.font = Font(
-            name='Calibri',
-            size=12,
-            bold=True,
-            color='FFFFFFFF'
-        )
-        header_style.fill = PatternFill(
-            start_color=self.colors['header_bg'],
-            end_color=self.colors['header_bg'],
-            fill_type='solid'
-        )
-        header_style.border = Border(
-            left=Side(style='thin', color='FF000000'),
-            right=Side(style='thin', color='FF000000'),
-            top=Side(style='thin', color='FF000000'),
-            bottom=Side(style='thick', color='FF000000')
-        )
-        header_style.alignment = Alignment(
-            horizontal='center',
-            vertical='center',
-            wrap_text=True
-        )
-        return header_style
-    
-    def create_data_style(self):
-        """Estilo para datos de tabla"""
-        data_style = NamedStyle(name="data_style")
-        data_style.font = Font(
-            name='Calibri',
-            size=11,
-            color='FF1F2937'
-        )
-        data_style.border = Border(
-            left=Side(style='thin', color='FFE5E7EB'),
-            right=Side(style='thin', color='FFE5E7EB'),
-            top=Side(style='thin', color='FFE5E7EB'),
-            bottom=Side(style='thin', color='FFE5E7EB')
-        )
-        data_style.alignment = Alignment(
-            horizontal='right',
-            vertical='center'
-        )
-        data_style.number_format = '$#,##0.00'
-        return data_style
-    
-    def create_period_style(self):
-        """Estilo para columna de períodos"""
-        period_style = NamedStyle(name="period_style")
-        period_style.font = Font(
-            name='Calibri',
-            size=11,
-            bold=True,
-            color='FF1E40AF'
-        )
-        period_style.fill = PatternFill(
-            start_color=self.colors['light'],
-            end_color=self.colors['light'],
-            fill_type='solid'
-        )
-        period_style.border = Border(
-            left=Side(style='thick', color='FF1E40AF'),
-            right=Side(style='thin', color='FFE5E7EB'),
-            top=Side(style='thin', color='FFE5E7EB'),
-            bottom=Side(style='thin', color='FFE5E7EB')
-        )
-        period_style.alignment = Alignment(
-            horizontal='center',
-            vertical='center'
-        )
-        return period_style
-
-class LicenseManager:
-    def __init__(self):
-        # URL de tu Firebase Realtime Database
-        self.firebase_url = "https://recibos-anses-default-rtdb.firebaseio.com"
-        self.machine_id = self.get_machine_id()
-    
-    def get_machine_id(self):
-        """Genera un ID único para la máquina"""
-        try:
-            # Combinar información del sistema para crear ID único
-            machine_info = f"{platform.node()}-{platform.machine()}-{getpass.getuser()}-{platform.platform()}"
-            return hashlib.sha256(machine_info.encode()).hexdigest()
-        except:
-            return "UNKNOWN_MACHINE"
-
-    def check_license(self):
-        """Verifica si esta máquina tiene licencia válida en Firebase"""
-        try:
-            # Verificar en Firebase
-            url = f"{self.firebase_url}/licenses/{self.machine_id}.json"
-            response = requests.get(url, timeout=10)
-            
-            if response.status_code == 200:
-                license_data = response.json()
-                
-                if license_data is None:
-                    return False, f"Máquina no registrada.\nID: {self.machine_id}"
-                
-                # Verificar si está activa
-                if license_data.get('active', False):
-                    # Verificar expiración si existe
-                    if 'expires_at' in license_data:
-                        try:
-                            expires_at = datetime.fromisoformat(license_data['expires_at'])
-                            if datetime.now() > expires_at:
-                                return False, "Licencia expirada"
-                        except:
-                            pass  # Si no se puede parsear la fecha, continuar
-                    
-                    # Actualizar último uso
-                    self.update_last_used()
-                    return True, "Licencia válida, cierre la ventana para continuar"
-                else:
-                    return False, f"Licencia desactivada.\nID: {self.machine_id}"
-            else:
-                return False, f"Error de conexión: {response.status_code}"
-                
-        except requests.exceptions.RequestException:
-            return False, f"Sin conexión a internet.\nID: {self.machine_id}"
-        except Exception as e:
-            return False, f"Error: {str(e)}"
-
-    def update_last_used(self):
-        """Actualiza la fecha de último uso en Firebase"""
-        try:
-            url = f"{self.firebase_url}/licenses/{self.machine_id}/last_used.json"
-            requests.put(url, json=datetime.now().isoformat(), timeout=5)
-        except:
-            pass  # No es crítico si falla
-
-class LicenseDialog(ctk.CTkToplevel):
-    def __init__(self, parent, license_manager):
-        super().__init__(parent)
-        self.parent = parent
-        self.license_manager = license_manager
-        self.license_valid = False
-        
-        # Variables para threading
-        self.license_result = None
-        self.checking_license = False
-        
-        self.setup_ui()
-        self.check_license()
-    
-    def setup_ui(self):
-        self.title("🔐 Verificación de Licencia")
-        self.geometry("600x500")
-        self.resizable(False, False)
-        
-        # Centrar ventana
-        self.transient(self.parent)
-        self.grab_set()
-        
-        # Frame principal
-        main_frame = ctk.CTkFrame(self)
-        main_frame.pack(fill="both", expand=True, padx=20, pady=20)
-        
-        # Título
-        title_label = ctk.CTkLabel(
-            main_frame,
-            text="🔐 VERIFICACIÓN DE LICENCIA",
-            font=ctk.CTkFont(size=20, weight="bold"),
-            text_color=["#1E40AF", "#60A5FA"]
-        )
-        title_label.pack(pady=(20, 10))
-        
-        # Información de la máquina
-        info_frame = ctk.CTkFrame(main_frame, fg_color=["#F8FAFC", "#1E293B"])
-        info_frame.pack(fill="x", pady=(0, 20))
-        
-        ctk.CTkLabel(
-            info_frame,
-            text="💻 INFORMACIÓN DE LA MÁQUINA",
-            font=ctk.CTkFont(size=16, weight="bold")
-        ).pack(pady=(15, 10))
-        
-        # Mostrar info de la máquina
-        machine_info = [
-            f"🖥️ Equipo: {platform.node()}",
-            f"👤 Usuario: {getpass.getuser()}",
-            f"🔧 Sistema: {platform.system()}",
-            f"🆔 ID: {self.license_manager.machine_id[:32]}..."
-        ]
-        
-        for info in machine_info:
-            ctk.CTkLabel(
-                info_frame,
-                text=info,
-                font=ctk.CTkFont(size=12),
-                text_color=["#6B7280", "#9CA3AF"]
-            ).pack(pady=2)
-        
-        ctk.CTkLabel(info_frame, text="", height=10).pack()
-        
-        # Estado de la licencia
-        self.status_frame = ctk.CTkFrame(main_frame, fg_color=["#FEF2F2", "#1F1F1F"])
-        self.status_frame.pack(fill="x", pady=(0, 20))
-        
-        self.status_label = ctk.CTkLabel(
-            self.status_frame,
-            text="🔄 Verificando licencia...",
-            font=ctk.CTkFont(size=14, weight="bold"),
-            text_color=["#DC2626", "#EF4444"]
-        )
-        self.status_label.pack(pady=15)
-        
-        # Instrucciones
-        instructions_frame = ctk.CTkFrame(main_frame, fg_color=["#F0F9FF", "#1E293B"])
-        instructions_frame.pack(fill="x", pady=(0, 20))
-        
-        ctk.CTkLabel(
-            instructions_frame,
-            text="📋 INSTRUCCIONES",
-            font=ctk.CTkFont(size=16, weight="bold")
-        ).pack(pady=(15, 10))
-        
-        instructions_text = ctk.CTkTextbox(instructions_frame, height=100, wrap="word")
-        instructions_text.pack(fill="x", padx=15, pady=(0, 15))
-        
-        instructions = """ℹ️ VERIFICACIÓN DE LICENCIA:
-Si tu máquina no está autorizada, contacta al administrador con tu ID de máquina.
-El administrador debe crear una licencia para tu máquina usando el generador de licencias.
-
-📞 CONTACTO: Proporciona tu ID de máquina completo al administrador."""
-        
-        instructions_text.insert("1.0", instructions)
-        instructions_text.configure(state="disabled")
-        
-        # Botones
-        buttons_frame = ctk.CTkFrame(main_frame, fg_color="transparent")
-        buttons_frame.pack(fill="x", pady=(0, 20))
-        
-        self.refresh_btn = ctk.CTkButton(
-            buttons_frame,
-            text="🔄 Verificar Nuevamente",
-            font=ctk.CTkFont(size=14, weight="bold"),
-            height=45,
-            command=self.check_license
-        )
-        self.refresh_btn.pack(side="left", fill="x", expand=True, padx=(0, 10))
-        
-        self.continue_btn = ctk.CTkButton(
-            buttons_frame,
-            text="✅ Continuar",
-            font=ctk.CTkFont(size=14, weight="bold"),
-            height=45,
-            state="disabled",
-            fg_color=["#10B981", "#059669"],
-            command=self.continue_app
-        )
-        self.continue_btn.pack(side="right", fill="x", expand=True, padx=(10, 0))
-        
-        # ID completo para copiar
-        id_frame = ctk.CTkFrame(main_frame, fg_color=["#F8FAFC", "#1E293B"])
-        id_frame.pack(fill="x")
-        
-        ctk.CTkLabel(
-            id_frame,
-            text="🆔 ID COMPLETO DE LA MÁQUINA:",
-            font=ctk.CTkFont(size=12, weight="bold")
-        ).pack(pady=(15, 5))
-        
-        self.id_textbox = ctk.CTkTextbox(id_frame, height=60, wrap="word")
-        self.id_textbox.pack(fill="x", padx=15, pady=(0, 15))
-        self.id_textbox.insert("1.0", self.license_manager.machine_id)
-        self.id_textbox.configure(state="disabled")
-    
-    def check_license(self):
-        """Verifica la licencia"""
-        self.refresh_btn.configure(text="🔄 Verificando...", state="disabled")
-        self.update_status(False, "🔄 Verificando licencia...")
-        
-        # Variables para comunicación entre threads
-        self.license_result = None
-        self.checking_license = True
-        
-        # Ejecutar verificación en hilo separado
-        thread = threading.Thread(target=self._check_license_thread)
-        thread.daemon = True
-        thread.start()
-        
-        # Iniciar polling para verificar resultados
-        self._poll_license_result()
-
-    def _check_license_thread(self):
-        """Hilo de verificación de licencia"""
-        try:
-            valid, message = self.license_manager.check_license()
-            # Guardar resultado en variable de instancia
-            self.license_result = (valid, message)
-            self.checking_license = False
-        except Exception as e:
-            # En caso de error en el hilo
-            self.license_result = (False, f"Error en verificación: {str(e)}")
-            self.checking_license = False
-
-    def _poll_license_result(self):
-        """Verifica periódicamente si hay resultados del hilo de licencia"""
-        if hasattr(self, 'checking_license') and not self.checking_license and self.license_result:
-            # Hay resultado disponible
-            valid, message = self.license_result
-            self._update_license_result(valid, message)
-            # Limpiar variables
-            self.license_result = None
-            self.checking_license = False
-        elif hasattr(self, 'checking_license') and self.checking_license:
-            # Aún verificando, programar siguiente verificación
-            self.after(100, self._poll_license_result)
-
-    def _update_license_result(self, valid, message):
-        """Actualiza el resultado de la verificación (ejecutado en hilo principal)"""
-        self.refresh_btn.configure(text="🔄 Verificar Nuevamente", state="normal")
-        self.update_status(valid, message)
-
-    def update_status(self, valid, message):
-        """Actualiza el estado visual de la licencia"""
-        self.license_valid = valid
-        
-        if valid:
-            self.status_frame.configure(fg_color=["#F0FDF4", "#1F2937"])
-            self.status_label.configure(
-                text=f"✅ {message}",
-                text_color=["#059669", "#10B981"]
-            )
-            self.continue_btn.configure(state="normal")
-        else:
-            self.status_frame.configure(fg_color=["#FEF2F2", "#1F1F1F"])
-            self.status_label.configure(
-                text=f"❌ {message}",
-                text_color=["#DC2626", "#EF4444"]
-            )
-            self.continue_btn.configure(state="disabled")
-
-    def continue_app(self):
-        """Continúa con la aplicación si la licencia es válida"""
-        if self.license_valid:
-            self.destroy()
-        else:
-            messagebox.showerror("Error", "Debes tener una licencia válida para continuar")
-
-class ConsoleWidget(ctk.CTkTextbox):
-    def __init__(self, parent, **kwargs):
-        super().__init__(parent, **kwargs)
-        self.configure(
-            font=ctk.CTkFont(family="Consolas", size=12),
-            text_color="#00FF00",
-            fg_color="#1a1a1a",
-            border_color="#333333",
-            border_width=2
-        )
-        self.log_count = 0
-    
-    def log(self, mensaje, tipo="info"):
-        timestamp = datetime.now().strftime("%H:%M:%S")
-        self.log_count += 1
-        
-        # Colores según tipo de mensaje
-        color_map = {
-            "info": "#00FF00",
-            "warning": "#FFA500", 
-            "error": "#FF4444",
-            "success": "#00FF88",
-            "process": "#00AAFF"
-        }
-        
-        color = color_map.get(tipo, "#00FF00")
-        
-        # Formato de consola profesional
-        line = f"[{timestamp}] [{self.log_count:03d}] {mensaje}\n"
-        
-        self.insert("end", line)
-        self.see("end")
-        
-        # Mantener solo las últimas 100 líneas
-        lines = self.get("1.0", "end").split('\n')
-        if len(lines) > 100:
-            self.delete("1.0", f"{len(lines)-100}.0")
 
 class ANSESDownloaderPro:
     def __init__(self):
